Merged PR 640512: [MerlinBot] - added auto PR comment to check if change should be applied on other embed repos
added auto PR comment to check if change should be applied on other embed repos.

When there is a change to a file configured for all of embed repos, we usually need to apply the change to all of the repos we maintain. For example - the following PR where we updated the signing profile:
https://dev.azure.com/powerbi/Embedded/_git/powerbi-javascript/commit/3c09aa92c8e1a4ffe7d081761eddc62b4e71b04b?refName=refs%2Fheads%2Fmaster&path=%2F.pipelines%2FOneBranch.Official.yml&_a=compare

Many times these changes are applied by ES team and this comment will help them to understand which repositories they need to change

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request adds an automated comment feature to check if changes should be applied to other embedded repositories.
- Added `/.config/merlinbot/nitpicker.yaml` to configure auto PR comments for validation of changes across embed repositories.