Merged PR 674975: Remove demos
Following SFI alerts, we were requested to fix a minor issue in the syntaxHighlighter.js files.

![image.png](https://powerbi.visualstudio.com/4c7b5adb-c2d0-4f18-b23c-edc4ac30f4e1/_apis/git/repositories/1dd63332-e003-46a3-ac6f-70213a1bcc2e/pullRequests/674975/attachments/image.png)

![image (2).png](https://powerbi.visualstudio.com/4c7b5adb-c2d0-4f18-b23c-edc4ac30f4e1/_apis/git/repositories/1dd63332-e003-46a3-ac6f-70213a1bcc2e/pullRequests/674975/attachments/image%20%282%29.png)

However, the demos are not maintained, and they are duplications of playground v1 codebase, which was deprecated and removed.

----
#### AI description  (iteration 1)
#### PR Classification
Code cleanup

#### PR Summary
This pull request removes all demo-related files and updates the `README.md` to reflect these changes.
- Deleted all files under the `demo` directory, including scripts, styles, images, and HTML files.
- Updated `README.md` to remove references to the deleted demo and provide links to alternative resources.